requirements.txt
# Core ML
numpy==1.26.4
pandas==2.2.2
scikit-learn==1.4.2
statsmodels==0.14.2
dowhy==0.11.1
torch==2.2.2
lifelines==0.29.0
anthropic==0.25.0
pyyaml==6.0.1
joblib==1.4.2
pydantic==2.7.0

# API Serving
fastapi==0.111.0
uvicorn==0.29.0

# Testing
pytest==8.2.0
hypothesis==6.100.0

# Code Quality
black==24.4.2
mypy==1.10.0
flake8==7.0.0

# Security
pip-audit==2.7.3
pre-commit==3.7.0
